+++ controller/teacher.js
@@ -1,7 +1,6 @@
 const { statusCode, successMessage, failMessage } = require('~/common/message');
 const { teacherService } = require('~/service');
-const { certificateService } = require('~/service');
-
+const { validate, teacherSchema } = require('~/validate');
 async function getAllTeachers(request, reply) {
   try {
     const data = await teacherService.getAllTeachers();
@@ -15,7 +14,12 @@ async function getTeacher(request, reply) {
   try {
     const { id } = request.params;
     const data = await teacherService.getTeacher(id);
-    return reply.status(statusCode.success).send({ data: data, message: successMessage.index });
+    if (data) {
+      return reply.status(statusCode.success).send({ data: data, message: successMessage.index });
+    }
+    else {
+      return reply.status(statusCode.notFound).send({ message: failMessage.notFoundID });
+    }
   } catch (err) {
     reply.status(statusCode.internalError).send({ message: failMessage.internalError });
   }
@@ -24,8 +28,14 @@ async function getTeacher(request, reply) {
 async function deleteTeacher(request, reply) {
   try {
     const { id } = request.params;
-    const data = await teacherService.deleteTeacher(id);
-    return reply.status(statusCode.success).send({ data: data, message: successMessage.index });
+
+    const Exist = await teacherService.getTeacher(id)
+    if (Exist) {
+      const data = await teacherService.deleteTeacher(id);
+      return reply.status(statusCode.success).send({ data: data, message: successMessage.index });
+    } else {
+      return reply.status(statusCode.notFound).send({ message: failMessage.notFoundID });
+    }
   } catch (err) {
     reply.status(statusCode.internalError).send({ message: failMessage.internalError });
   }
@@ -33,8 +43,24 @@ async function deleteTeacher(request, reply) {
 
 async function insertTeacher(request, reply) {
   try {
-    const data_teacher = request.body;
-    const data = await teacherService.insertTeacher(data_teacher);
+    const body = request.body;
+    const check = validate(body, teacherSchema.insertTeacherSchema, reply);
+    if (check === false) {
+      return reply.status(statusCode.badRequest).send({ message: failMessage.invalidData });
+    };
+    const dataInsert = {
+      name: body.name,
+      avatar: body.avatar,
+      introduction: body.introduction,
+      nationality: body.nationality,
+      experiences: body.experiences,
+      certificates: body.certificates,
+    }
+    const Exist = await teacherService.checkTeacher(dataInsert);
+    if (Exist) {
+      return reply.status(statusCode.badRequest).send({ message: failMessage.existData });
+    }
+    const data = await teacherService.insertTeacher(dataInsert);
     return reply.status(statusCode.success).send({ data: data, message: successMessage.index });
   } catch (err) {
     reply.status(statusCode.internalError).send({ message: failMessage.internalError });
@@ -43,10 +69,24 @@ async function insertTeacher(request, reply) {
 
 async function updateTeacher(request, reply) {
   try {
-    const data_teacher = request.body;
+    const body = request.body;
     const { id } = request.params;
-    const data = await teacherService.updateTeacher({id, data_teacher});
-    return reply.status(statusCode.success).send({ data: data, message: successMessage.index });
+    const Exist = await teacherService.getTeacher(id)
+    if (Exist) {
+      const check = validate(body, teacherSchema.updateTeacherSchema, reply);
+      if (check === false) {
+        return reply.status(statusCode.badRequest).send({ message: failMessage.invalidData });
+      };
+      if (!body) {
+        return reply.status(statusCode.badRequest).send({ message: failMessage.emptyData });
+      }
+      const dataUpdate = body;
+      const data = await teacherService.updateTeacher({ id, dataUpdate });
+      return reply.status(statusCode.success).send({ data: data, message: successMessage.index });
+    }
+    else {
+      return reply.status(statusCode.notFound).send({ message: failMessage.notFoundID });
+    }
   } catch (err) {
     reply.status(statusCode.internalError).send({ message: failMessage.internalError });
   }
